debug

+++ DataProcess.py
@@ -46,7 +46,7 @@ def prepare_data(path):
                 sub_img = img[x_start:x_end, y_start:y_end, :]
                 data.append(sub_img)
 
-    data = numpy.array(data)  #list has no shape
+    data = numpy.array(data, dtype= numpy.float32)  #list has no shape
     print 'data.shape:',data.shape          
     return data
 
@@ -55,8 +55,8 @@ def write_hdf5(data, label,output_filename):
     This function is used to save image data or its label(s) to hdf5 file.
     output_file.h5,contain data and label
     """
-    x = data.astype(numpy.int)
-    y = label.astype(numpy.int)
+    x = data.astype(numpy.float32)
+    y = label.astype(numpy.float32)
     with h5py.File(output_filename, 'w') as h:
         h.create_dataset('data', data=x, shape=x.shape)
         h.create_dataset('label', data=y, shape=y.shape)
